Migrate existing posts from prior resources page
