+++ frontend/content/posts/contribute-to-the-aws-cdk/index.yml
@@ -1,0 +1,9 @@
+title: "Contribute to the AWS CDK"
+excerpt: "Ever wondering how to create a pull request for AWS CDK? In this video, we'll walk you through all the primary steps to submit your first AWS CDK pull request"
+authorIds:
+  - pahud-hsieh
+url: https://youtu.be/OXQSSibrt-A
+createdAt: "2020-09-18 19:00"
+categories:
+  - AWS-CDK
+  - Video
