+++ frontend/content/posts/enhanced-cicd-aws-cdk/index.yml
@@ -1,0 +1,11 @@
+title: "Enhanced CI/CD with AWS CDK"
+excerpt: "AWS Online Tech Talk about CI/CD pipelines in AWS CDK with live demo and example code repo"
+authorIds:
+  - adam-keller
+url: https://youtu.be/1ps0Wh19MHQ
+createdAt: "2020-09-18 19:00"
+categories:
+  - AWS-CDK
+  - Video
+  - Python
+  - CI/CD
