+++ frontend/content/posts/lets-checkout-cdk-pipelines/index.yml
@@ -1,0 +1,10 @@
+title: "Let's Check out CDK Pipelines"
+excerpt: "Live coding a CDK CI/CD pipeline"
+authorIds:
+  - darko-mesaros
+url: https://www.twitch.tv/videos/737014680
+createdAt: "2020-09-18 19:00"
+categories:
+  - AWS-CDK
+  - CDK Pipelines
+  - Video
