+++ frontend/content/users/adam-keller/index.yml
@@ -1,0 +1,4 @@
+name: Adam Keller
+avatar: avatar.png
+twitter: realadamjkeller
+createdAt: "2020-09-17 18:00"
