+++ frontend/content/users/christian-weber/index.yml
@@ -1,0 +1,4 @@
+name: Christian Weber
+avatar: avatar.jpg
+twitter: classicwebdog
+createdAt: "2020-09-17 18:00"
