+++ frontend/content/users/darko-mesaros/index.yml
@@ -1,0 +1,4 @@
+name: Darko Mesaroš
+avatar: avatar.jpg
+twitter: darkosubotica
+createdAt: "2020-09-17 18:00"
