+++ frontend/content/users/matthew-bonig/index.yml
@@ -1,0 +1,4 @@
+name: Matthew Bonig
+avatar: avatar.jpg
+twitter: mattbonig
+createdAt: "2020-09-17 18:00"
