+++ frontend/content/users/rico-huijbers/index.yml
@@ -1,0 +1,4 @@
+name: Rico Huijbers
+avatar: avatar.jpeg
+twitter: rix0rrr
+createdAt: "2020-09-17 18:00"
